Extract use ID calculation into helper function
This allows us to re-use it across the new and old website.
